test: split bounded e2e process support
